Update patch set 8
Patch Set 8:

> Patch Set 7:
> 
> one point of confusion that still remains: if the calls like set, find_and_get are atomic, under what scenarios does a user need to aquire a lock? 
> 
> Probably needs a docs/user guide

A user might need a distributed 'mutex' kind of locking mechanism for some other shared resource that is not necessary SDL itself. Some time back we implemented SDL locking service in Golang and that time (if I remember this correctly)  the reason why locking service implementation was asked from us was the "big red button" implementation in Nodeb-rnib. However it seems that currently no one is actually using locking service but the SDL locking implementation in Golang is there ready to take into use if needed and now the similar implementation will be available for SDL in python too.

Patch-set: 8
Label: Code-Review=0