Update patch set 2
Patch Set 2:

Yes, because Tommy needs your component next month AND we'd like to test the full process, please do bump your version to 1.0.1, add a change set with that as a separate change, wait for it to verify, then merge it. When that is all done, then bump the version number listed in this release yaml file, wait for the verify etc.

Patch-set: 2