Implement regular expression support for key searching
There is a use case to use only some part of the full key as a
searching criterion, as an example of such a criterion could be gNBID,
arp or qci. That's why change find_keys() and find_and_get() API
functions to take in a key search pattern instead of a key prefix.
A pattern can contain glob-style regular expression to match a key.

Removed also 'atomic' parameter of find_and_get(), because only
reasonable selection would be to use non-atomic option, which does not
block the whole Redis for the long time like atomic option does.

Signed-off-by: Timo Tietavainen <timo.tietavainen@nokia.com>
Change-Id: I870089f51b1ce4d72c7984220af0f770b562563f