Update patch set 1
Patch Set 1:

Oh I see now.  The A1 repo has all the files at the root.  The SDLPY repo has the files below ricsdl-package.  So that's DEFINITELY worth a try because maven runs in the project root, but the coverage.xml is written to the tox dir

Oh this awkward marriage of technologies makes work forever.

Patch-set: 1